LayoutTests/fast/dom/HTMLScriptElement/script-for-attribute-unexpected-execution-expected.txt

If a script has a for attribute, then it was intended to only be run under certain conditions, often as a result of a certain window event.
Since we don't yet support the full for attribute syntax we would run these scripts as we parsed them, often causing unintentional breakage of the site in question.
You should *not* see any failure when running this test. If you do, we're not properly running these scripts only when they were intended to be run.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS scriptForExecuted is false
PASS scriptForExecuted is false
PASS scriptForExecuted is false
PASS successfullyParsed is true

TEST COMPLETE
